+++ src/Ydb.Sdk/tests/Ado/YdbDataReaderTests.cs
@@ -20,8 +20,6 @@ public class YdbDataReaderTests
 
         Assert.Equal("No row is available", Assert.Throws<InvalidOperationException>(() => reader.GetValue(0)).Message);
 
-        Assert.True(reader.NextResult());
-
         Assert.Equal("No row is available",
             Assert.Throws<InvalidOperationException>(() => reader.GetValue(0)).Message); // Need Read()
 
@@ -32,11 +30,18 @@ public class YdbDataReaderTests
             Assert.Throws<IndexOutOfRangeException>(() => reader.GetBoolean(1)).Message);
 
         Assert.False(reader.Read());
-        Assert.True(reader.IsClosed);
+        Assert.False(reader.IsClosed);
 
-        Assert.Equal("The reader is closed",
+        Assert.Equal("No row is available",
             Assert.Throws<InvalidOperationException>(() => reader.GetValue(0)).Message);
         Assert.Empty(statuses);
+
+        await reader.CloseAsync();
+        Assert.True(reader.IsClosed);
+        Assert.Equal("The reader is closed",
+            Assert.Throws<InvalidOperationException>(() => reader.GetValue(0)).Message);
+        Assert.Equal("The reader is closed",
+            Assert.Throws<InvalidOperationException>(() => reader.Read()).Message);
     }
 
     [Fact]
@@ -56,7 +61,6 @@ public class YdbDataReaderTests
         var statuses = new List<Status>();
         var reader = await YdbDataReader.CreateYdbDataReader(EnumeratorSuccess(2), statuses.Add);
 
-        Assert.True(reader.NextResult());
         Assert.True(reader.NextResult());
         Assert.True(reader.Read());
         Assert.True((bool)reader.GetValue(0));